:wrench: rename store and router file
