+++ src/main.js
@@ -1,8 +1,8 @@
 import Vue from "vue";
 import App from "./App.vue";
-import router from "./router";
-import store from "./store";
 import vuetify from "./plugins/vuetify";
+import router from "./router/router.js";
+import store from "./store/store.js";
 
 Vue.config.productionTip = false;
